listening-comp/requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.2.0
numpy==1.24.0

# AWS
boto3==1.34.0

# Audio Processing
openai-whisper==20231117
ffmpeg-python==0.2.0

# YouTube Integration
youtube-transcript-api==0.6.2

# Vector Storage
chromadb==0.4.22

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0